omg yesss GIFs are playing
The bug was that simply using the assignment operator
with a Mat object only copies the header/pointer, not the data itself.

That's why only the "last" image of a gif was ever playing, because the
data contained in the list of images only ever pointed to the same data,
i.e. the last image read.

TODO: uh, so why are like the first two pixels corrupted between reading
the image and returning a vector?